refactor(git): extract discard into its own module
The discard rework pushed status.ts past the thousand-line architecture
budget. Move discard and its HEAD probe to server/git/discard.ts beside
the other per-operation modules; the code moves verbatim and status.ts
re-exports it from its operations factory unchanged.
